auto conversion of DataValue to/from python types
git-svn-id: https://open-ms.svn.sourceforge.net/svnroot/open-ms/OpenMS@11088 6adb6e08-d915-0410-941f-83917bcadc18

+++ pyOpenMS/tests/unittests/test_MRMFeatureFinderScoring.py
@@ -41,8 +41,8 @@ class TestMRMFeatureFinderScoring(unittest.TestCase):
         self.assertAlmostEqual(output.size(), 3)
         self.assertAlmostEqual(output[0].getRT(), 3119.092041015)
         self.assertAlmostEqual(output[0].getIntensity(), 3574.232421875)
-        self.assertAlmostEqual(output[0].getMetaValue("var_xcorr_shape_weighted").toDouble(), 0.997577965259552)
-        self.assertAlmostEqual(output[0].getMetaValue("sn_ratio").toDouble(), 86.00413513183594)
+        self.assertAlmostEqual(output[0].getMetaValue("var_xcorr_shape_weighted"), 0.997577965259552)
+        self.assertAlmostEqual(output[0].getMetaValue("sn_ratio"), 86.00413513183594)
 
 if __name__ == '__main__':
     unittest.main()
